# xhs_backend/api_docs/openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "小红书评论维护系统",
    "description": "小红书评论数据的管理和维护系统",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/system/status": {
      "get": {
        "tags": [
          "系统"
        ],
        "summary": "System Status",
        "description": "获取系统状态信息\n\nArgs:\n    current_user: 当前用户\n    \nReturns:\n    系统状态信息",
        "operationId": "system_status_api_system_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response System Status Api System Status Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/system/database-stats": {
      "get": {
        "tags": [
          "系统"
        ],
        "summary": "Database Stats",
        "description": "获取数据库统计信息\n\nArgs:\n    current_user: 当前用户\n    \nReturns:\n    数据库统计信息",
        "operationId": "database_stats_api_system_database_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Database Stats Api System Database Stats Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/login"
          }
        }
      }
    }
  }
}